refactor(bridge): extract completion postlude
